добавлены зависимости для использования retrofit, раскидал классы по пакетам по-другому
