+++ app/build.gradle
@@ -66,7 +66,8 @@ dependencies {
     implementation "androidx.navigation:navigation-dynamic-features-fragment:$nav_version"
     androidTestImplementation "androidx.navigation:navigation-testing:$nav_version"
     implementation "androidx.navigation:navigation-compose:2.4.0-alpha05"
-    implementation 'com.google.code.gson:gson:2.8.6'
+    implementation 'com.squareup.retrofit2:2.9.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
 }
 repositories {
     google()
