--- app/src/main/java/project/paveltoy/movietap/data/Sections.kt
@@ -1,4 +1,0 @@
-package project.paveltoy.movietap.data
-
-interface Sections {
-}
