+++ app/src/main/java/project/paveltoy/movietap/data/entity/Section.kt
@@ -1,4 +1,4 @@
-package project.paveltoy.movietap.data
+package project.paveltoy.movietap.data.entity
 
 data class Section(
     val id: Int,
