+++ app/src/main/java/project/paveltoy/movietap/data/entity/Sections.kt
@@ -1,0 +1,4 @@
+package project.paveltoy.movietap.data.entity
+
+interface Sections {
+}
