+++ app/src/main/java/project/paveltoy/movietap/ui/MainViewModel.kt
@@ -1,8 +1,11 @@
-package project.paveltoy.movietap.viewmodels
+package project.paveltoy.movietap.ui
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import project.paveltoy.movietap.data.*
+import project.paveltoy.movietap.data.entity.MovieEntity
+import project.paveltoy.movietap.data.entity.Movies
+import project.paveltoy.movietap.data.repository.MovieRepo
+import project.paveltoy.movietap.data.repository.TMDBMovieRepo
 
 class MainViewModel : ViewModel() {
     val clickedMovieLiveData = MutableLiveData<MovieEntity>()
