+++ app/src/main/java/project/paveltoy/movietap/ui/main/MainFragment.kt
@@ -1,4 +1,4 @@
-package project.paveltoy.movietap.ui
+package project.paveltoy.movietap.ui.main
 
 import android.content.Context
 import android.os.Bundle
@@ -8,29 +8,26 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
-import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.snackbar.BaseTransientBottomBar
 import com.google.android.material.snackbar.Snackbar
 import project.paveltoy.movietap.R
-import project.paveltoy.movietap.data.MovieEntity
 import project.paveltoy.movietap.data.getTextForIsFavoriteSnackbar
-import project.paveltoy.movietap.databinding.FragmentFavoriteMovieBinding
-import project.paveltoy.movietap.ui.adapters.FavoritesAdapter
-import project.paveltoy.movietap.viewmodels.MainViewModel
+import project.paveltoy.movietap.databinding.FragmentMainBinding
+import project.paveltoy.movietap.ui.MainViewModel
 
-class FavoriteMovieFragment : Fragment() {
-    private var _binding: FragmentFavoriteMovieBinding? = null
-    val binding get() = _binding!!
-    private lateinit var favoriteRecyclerView: RecyclerView
-    private lateinit var adapter: FavoritesAdapter
+class MainFragment : Fragment() {
+    private var _binding: FragmentMainBinding? = null
+    private val binding get() = _binding!!
+    private lateinit var mainRecyclerView: RecyclerView
     private lateinit var viewModel: MainViewModel
-    private lateinit var data: List<MovieEntity>
+    private lateinit var movieSectionsAdapter: MovieSectionsAdapter
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
         viewModel = ViewModelProvider(requireActivity()).get(MainViewModel::class.java)
-        adapter = FavoritesAdapter()
+        movieSectionsAdapter = MovieSectionsAdapter()
     }
 
     override fun onCreateView(
@@ -38,47 +35,45 @@ class FavoriteMovieFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentFavoriteMovieBinding.inflate(inflater, container, false)
+        _binding = FragmentMainBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        serRecyclerView()
+        setRecyclerView()
         setAdapter()
+        viewModel.moviesLiveData.observe(viewLifecycleOwner,
+            { movieSectionsAdapter.notifyDataSetChanged() })
     }
 
-    private fun serRecyclerView() {
-        favoriteRecyclerView = binding.favoriteRecyclerView
-        favoriteRecyclerView.layoutManager = GridLayoutManager(context, 2)
-        favoriteRecyclerView.adapter = adapter
+    private fun setRecyclerView() {
+        mainRecyclerView = binding.verticalRecyclerView
+        mainRecyclerView.layoutManager = LinearLayoutManager(context)
+        mainRecyclerView.adapter = movieSectionsAdapter
     }
 
     private fun setAdapter() {
-        data = viewModel.getFavoriteMovies()
+        movieSectionsAdapter.data = viewModel.getMovies()
+        movieSectionsAdapter.notifyDataSetChanged()
         setOnItemClickListener()
         setOnFavoriteChanged()
-        adapter.data = data
-        adapter.notifyDataSetChanged()
     }
 
     private fun setOnItemClickListener() {
-        adapter.onItemClick = {
+        movieSectionsAdapter.onItemClick = {
             viewModel.clickedMovieLiveData.value = it
-            findNavController().navigate(R.id.action_favorite_movie_fragment_to_detailFragment)
+            findNavController().navigate(R.id.action_mainFragment_to_detailFragment)
         }
     }
 
     private fun setOnFavoriteChanged() {
-        adapter.onFavoriteChanged = {movie: MovieEntity, index: Int ->
-            movie.isFavorite = !movie.isFavorite
-            viewModel.clickedMovieLiveData.value = movie
-            val text = getTextForIsFavoriteSnackbar(resources, movie.title, movie.isFavorite)
+        movieSectionsAdapter.onFavoriteChanged = {
+            it.isFavorite = !it.isFavorite
+            val text = getTextForIsFavoriteSnackbar(resources, it.title, it.isFavorite)
             Snackbar.make(requireView(), text, BaseTransientBottomBar.LENGTH_LONG)
                 .setAnchorView(R.id.bottom_navigation)
                 .show()
-            adapter.data = viewModel.getFavoriteMovies()
-            adapter.notifyItemRemoved(index)
         }
     }
