+++ app/src/main/java/project/paveltoy/movietap/ui/main/MovieAdapter.kt
@@ -1,9 +1,9 @@
-package project.paveltoy.movietap.ui.adapters
+package project.paveltoy.movietap.ui.main
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import project.paveltoy.movietap.data.MovieEntity
+import project.paveltoy.movietap.data.entity.MovieEntity
 import project.paveltoy.movietap.databinding.ItemMovieBinding
 
 class MovieAdapter : RecyclerView.Adapter<MovieAdapter.BaseViewHolder>() {
